requirements.txt
google-generativeai
chromadb
python-dotenv
streamlit
pandas
plotly
pyyaml